impr(claims): address PR comments
